hypot

+++ src/io/metadevs/vnazarov/Main.java
@@ -8,9 +8,9 @@ public class Main {
         System.out.println("Программа считает гипотинузу треугольника при помощи двух катетов.");
         Scanner sc = new Scanner(System.in);
         System.out.println("Введите катет 1:");
-        double side1 = sc.nextInt();
+        double side1 = sc.nextDouble();
         System.out.println("Введите катет 2:");
-        double side2 = sc.nextInt();
+        double side2 = sc.nextDouble();
         sc.close();
         double hyp = Math.hypot(side1, side2);
         System.out.print("Гипотенуза треугольника с катетами ");
